Command Integration with Smart Dog
This project enables a smart dog to perceive voice commands coming from the user and direct them. Thanks to this system developed using Python and artificial intelligence technologies, the dog moves in accordance with the specified commands. It is a complete solution that combines voice recognition, neural network model and motion control modules.